# SCA fixes for BN in EC