This PR addresses 2 related but independent side-channel issues
## Fix a SCA leak using `BN_bn2bin()`

`BN_bn2bin()` is not constant-time and leaks the number of bits in the processed `BIGNUM`.

The specialized methods in `ecp_nistp224.c`, `ecp_nistp256.c` and `ecp_nistp521.c` internally used `BN_bn2bin()` to convert scalars into the internal fixed length representation.

This can leak during ECDSA/ECDH key generation or handling the nonce while generating an ECDSA signature, when using these implementations.  The amount and risk of leaked information useful for a SCA attack varies for each of the three curves, as it depends mainly on the ratio between the bitlength of the curve subgroup order (governing the size of the secret nonce/key) and the limb size for the internal `BIGNUM` representation (which depends on the compilation target architecture).

To fix this, we replace `BN_bn2bin()` with `BN_bn2binpad()`, bounding the output length to the width of the internal representation buffer: this length is public.

Internally the final implementation of both `BN_bn2binpad()` and `BN_bn2bin()` already has masking in place to avoid leaking `bn->top` through memory access patterns.
Memory access pattern still leaks `bn->dmax`, the size of the lazily allocated buffer for representing the `BIGNUM`, which is inevitable with the current `BIGNUM` architecture: reading past `bn->dmax` would be an out-of-bound read.
As such, it's the caller responsibility to ensure that `bn->dmax` does not leak secret information, by explicitly expanding the internal `BIGNUM` buffer to a public value sufficient to avoid any lazy reallocation while manipulating it: this is already done at the top level alongside setting the `BN_FLG_CONSTTIME`.

Finally, the internal implementation of `BN_bn2binpad()` indirectly calls `BN_num_bits()` via `BN_num_bytes()`: the current implementation of `BN_num_bits()` can leak information to a SCA attacker, and is addressed in the next commit.

## Make `BN_num_bits()` consttime upon `BN_FLG_CONSTTIME`

This issue was partially addressed by commit 972c87dfc7e765bd28a4964519c362f0d3a58ca4, which hardened its callee `BN_num_bits_word()` to avoid leaking the most-significant word of its argument via branching and memory access pattern.
The commit message also reported:
> There are a few places where `BN_num_bits` is called on an input where
> the bit length is also secret. This does *not* fully resolve those
> cases as we still only look at the top word.

`BN_num_bits()` is called directly or indirectly (e.g., through `BN_num_bytes()` or `BN_bn2binpad()` ) in various parts of the `crypto/ec` code, notably in all the currently supported implementations of scalar multiplication (in the generic path through `ec_scalar_mul_ladder()` as well as in dedicated methods like `ecp_nistp{224,256,521}.c` and `ecp_nistz256.c`).

Under the right conditions, a motivated SCA attacker could retrieve the secret bitlength of a secret nonce through this vulnerability, potentially leading, ultimately, to recover a long-term secret key.

With this commit, exclusively for `BIGNUM`s that are flagged with `BN_FLG_CONSTTIME`, instead of accessing only `bn->top`, all the limbs of the `BIGNUM` are accessed up to `bn->dmax` and bitwise masking is used to avoid branching.

Memory access pattern still leaks `bn->dmax`, the size of the lazily allocated buffer for representing the `BIGNUM`, which is inevitable with the current `BIGNUM` architecture: reading past `bn->dmax` would be an out-of-bound read.
As such, it's the caller responsibility to ensure that `bn->dmax` does not leak secret information, by explicitly expanding the internal `BIGNUM` buffer to a public value sufficient to avoid any lazy reallocation while manipulating it: this should be already done at the top level alongside setting the `BN_FLG_CONSTTIME`.

# Benchmarks

I ran some benchmarks to make sure the second fix does not have an excessively negative effect on performance: the results suggest that the performance loss is negligible for RSA, DSA and ECDSA (ecp_smpl (p384), ec2_smpl (b283), ecp_nistp224, ecp_nistz256)

## Before

```
version: 3.0.0-dev
built on: built on: Thu Aug  1 23:56:53 2019 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr)
compiler: ccache clang-3.9 -fPIC -pthread -m64 -Wa,--noexecstack -Qunused-arguments -Wall -O3 -g -DDEBUG_UNUSED -DPEDANTIC -pedantic -Wno-long-long -Wall -Wextra -Wno-unused-parameter -Wno-missing-field-initializers -Wswitch -Wsign-compare -Wshadow -Wformat -Wtype-limits -Wundef -Werror -Wmissing-prototypes -Wstrict-prototypes -Wswitch-default -Wno-parentheses-equality -Wno-language-extension-token -Wno-extended-offsetof -Wconditional-uninitialized -Wincompatible-pointer-types-discards-qualifiers -Wno-unknown-warning-option -Wmissing-variable-declarations -Qunused-arguments -Wno-string-plus-int -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DNDEBUG
                  sign    verify    sign/s verify/s
rsa 2048 bits 0.000601s 0.000018s   1664.4  56525.3
                  sign    verify    sign/s verify/s
dsa 2048 bits 0.000263s 0.000221s   3804.4   4534.6
                              sign    verify    sign/s verify/s
 224 bits ecdsa (nistp224)   0.0000s   0.0001s  21518.9  11205.1
 256 bits ecdsa (nistp256)   0.0000s   0.0001s  40744.8  14774.5
 384 bits ecdsa (nistp384)   0.0010s   0.0008s    980.0   1316.3
 283 bits ecdsa (nistb283)   0.0006s   0.0011s   1721.0    881.4
```

## After

```
version: 3.0.0-dev
built on: built on: Fri Aug  2 00:28:07 2019 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr)
compiler: ccache clang-3.9 -fPIC -pthread -m64 -Wa,--noexecstack -Qunused-arguments -Wall -O3 -g -DDEBUG_UNUSED -DPEDANTIC -pedantic -Wno-long-long -Wall -Wextra -Wno-unused-parameter -Wno-missing-field-initializers -Wswitch -Wsign-compare -Wshadow -Wformat -Wtype-limits -Wundef -Werror -Wmissing-prototypes -Wstrict-prototypes -Wswitch-default -Wno-parentheses-equality -Wno-language-extension-token -Wno-extended-offsetof -Wconditional-uninitialized -Wincompatible-pointer-types-discards-qualifiers -Wno-unknown-warning-option -Wmissing-variable-declarations -Qunused-arguments -Wno-string-plus-int -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DNDEBUG
                  sign    verify    sign/s verify/s
rsa 2048 bits 0.000603s 0.000018s   1657.9  56646.6
                  sign    verify    sign/s verify/s
dsa 2048 bits 0.000262s 0.000224s   3810.3   4454.5
                              sign    verify    sign/s verify/s
 224 bits ecdsa (nistp224)   0.0000s   0.0001s  21348.0  11186.0
 256 bits ecdsa (nistp256)   0.0000s   0.0001s  40201.4  14706.5
 384 bits ecdsa (nistp384)   0.0010s   0.0008s    974.0   1316.5
 283 bits ecdsa (nistb283)   0.0006s   0.0012s   1693.5    866.4
```

# Acknowledgments

Thanks to David Schrammel and Samuel Weiser for reporting this issue through responsible disclosure.

# Responsible disclosure

The initial security assessment evaluated the severity of this issue to be quite low, hence the fix is released with a public PR.