@romen 2019-08-02
## Backporting

I haven't tested it yet, but this should cherry-pick cleanly to 1.1.1, so I would ask the reviewers to approve for 1.1.1 as well.

Being a security issue, should this backported to 1.0.2 and 1.1.0 as well?