@davidben 2019-08-02
#6640 strikes again! :-)

Switching from `BN_bn2bin` to `BN_bn2binpad` is a good plan regardless. I'm dubious of the `BN_num_bits` change. That OpenSSL needs it in `BN_bn2binpad` is a symptom of #6640. 972c87d intentionally did not do this because, as I wrote, "those cases should not call BN_num_bits at all". I'm not aware of any crypto primitive whose correct implementation would require such an operation.

That's not to say you shouldn't add bandaids and partial workarounds where they make sense, but I don't think this is the right bandaid. Adding new dependencies on `BN_FLG_CONSTTIME` is undesirable, especially one that turns an O(1) operation to O(N). See #8865. And as this is only needed for a bandaid, you shouldn't entrench it in a public API. 

Instead, I would recommend just having `BN_bn2binpad` do this internally, at which point you don't even need the full complexity of the bit computation. We did roughly this in our very early days. (We didn't figure out how `BIGNUM` could be saved until years later.)
https://boringssl.googlesource.com/boringssl/+/6887edb917ba9ecbed85f9d63ec36638b1d1dbb6%5E%21/

Note, however, this is a *partial* workaround. It does *not* fix the root issue. The computation that filled in `bn->top` already leaked it. To fix the actual leak, you need to address #6640. The issue description has a roadmap. Once you do that, you can retire the bandaid, as we did:
https://boringssl.googlesource.com/boringssl/+/3b3e12d81ee52ee293b4a2dc4f416b209188e744%5E%21/