@sam1013 2019-08-02
> I'm dubious of the `BN_num_bits` change. That OpenSSL needs it in `BN_bn2binpad` is a symptom of #6640. [972c87d](https://github.com/openssl/openssl/commit/972c87dfc7e765bd28a4964519c362f0d3a58ca4) intentionally did not do this because, as I wrote, "those cases should not call BN_num_bits at all". I'm not aware of any crypto primitive whose correct implementation would require computing a secret bit length.
> 
> That's not to say you shouldn't add bandaids and partial workarounds where they make sense, but I don't think this is the right bandaid. Adding new dependencies on `BN_FLG_CONSTTIME` is undesirable, especially one that turns an O(1) operation to O(N). See #6078. [edit: fixed the bug link... GitHub completion] And as this is only needed for a bandaid, you shouldn't entrench it in a public API.

As you mentioned, one should not call `BN_num_bits` at all on private input. So the O(N) slowdown should only affect vulnerable code shown below. Note that `BN_num_bits`is not only called in `BN_bn2binpad` but also in
[ec_scalar_mul_ladder](https://github.com/openssl/openssl/blob/3051bf2afab7ac8b7b9c64e68755d1addd2fb8ff/crypto/ec/ec_mult.c#L222)
[ec_GFp_nistp521_points_mul](https://github.com/openssl/openssl/blob/c2969ff6e70b10f71fbd97c1d0b0cffc92bd69df/crypto/ec/ecp_nistp521.c#L1961)
[ec_GFp_nistp521_points_mul](https://github.com/openssl/openssl/blob/c2969ff6e70b10f71fbd97c1d0b0cffc92bd69df/crypto/ec/ecp_nistp521.c#L2008)

More instances are below, although they are probably not exploitable since all of them are on a word boundary:
[ec_GFp_nistp224_points_mul](https://github.com/openssl/openssl/blob/72962d025f875ac35114ac090b878ee18b246144/crypto/ec/ecp_nistp224.c#L1496)
[ec_GFp_nistp224_points_mul](https://github.com/openssl/openssl/blob/72962d025f875ac35114ac090b878ee18b246144/crypto/ec/ecp_nistp224.c#L1543)
[ecp_nistz256_windowed_mul](https://github.com/openssl/openssl/blob/15972296af6b98ae495ada9d4695f2a0e71f891c/crypto/ec/ecp_nistz256.c#L614)
[ecp_nistz256_windowed_mul](https://github.com/openssl/openssl/blob/15972296af6b98ae495ada9d4695f2a0e71f891c/crypto/ec/ecp_nistz256.c#L1192)
[ec_GFp_nistp256_points_mul](https://github.com/openssl/openssl/blob/8f58ede09572dcc6a7e6c01280dd348240199568/crypto/ec/ecp_nistp256.c#L2121)
[ec_GFp_nistp256_points_mul](https://github.com/openssl/openssl/blob/8f58ede09572dcc6a7e6c01280dd348240199568/crypto/ec/ecp_nistp256.c#L2170)
[rsa_sp800_56b_derive_params_from_pq](https://github.com/openssl/openssl/blob/952abb152189715f7b035d5446af947ec41e8a4b/crypto/rsa/rsa_sp800_56b_gen.c#L212)

> Note, however, this is a _partial_ workaround. It does _not_ fix the root issue. 

Agree.

> The computation that filled in `bn->top` already leaked it. 

Not neessarily. `bn->top` leaks the word length, `BN_bn2bin` leaks the byte length.