@romen 2019-08-02
> I'm dubious of the `BN_num_bits` change. That OpenSSL needs it in `BN_bn2binpad` is a symptom of #6640. [972c87d](https://github.com/openssl/openssl/commit/972c87dfc7e765bd28a4964519c362f0d3a58ca4) intentionally did not do this because, as I wrote, "those cases should not call BN_num_bits at all". I'm not aware of any crypto primitive whose correct implementation would require computing a secret bit length.
> 
> That's not to say you shouldn't add bandaids and partial workarounds where they make sense, but I don't think this is the right bandaid. Adding new dependencies on `BN_FLG_CONSTTIME` is undesirable, especially one that turns an O(1) operation to O(N). See #6078. [edit: fixed the bug link... GitHub completion] And as this is only needed for a bandaid, you shouldn't entrench it in a public API.

@sam1013 listed the other places where `BN_num_bits()` is called directly rather than indirectly.

I would say that the need to pad the secret scalar by adding once or twice the modulus is a legitimate use for a constant time `BN_num_bits()`: we can do that either conditional to the BN_FLG_CONSTTIME or adding a separate `BN_num_bits_consttime()`.

@davidben , maybe for these use cases you would prefer a separate dedicate function to prevent branch prediction attacks on the check for the flag inside `BN_num_bits()`?
That seems sensible to me, and I would gladly implement that change, a downside is that it will probably make backporting more cumbersome?

Addressing all the places where we do `if (bn->flags & BN_FLG_CONSTTIME)` to replace them with dedicated functions, instead feels like a completely separate PR that would overlap with the ongoing discussions on how to refactor the `BIGNUM` layer.