@romen 2019-08-02
I went ahead and pushed another fixup commit that separates `BN_num_bits_consttime()` into its own (`static inline`) function.

This seems cleaner anyway and gives us more flexibility if we decide that in some branches we want to make `BN_num_bits_consttime()` `libcrypto`-internal/public and replace calls to `BN_num_bits()` with explicit `consttime` ones.